feat: reorganize alternatif components; move to dedicated directory and update styles
