+++ src/App.jsx
@@ -3,7 +3,7 @@ import LoginPage from "@/pages/auth/loginPage";
 import DashboardPage from "@/pages/dashboard";
 import KriteriaList from "@/pages/kriteria/list";
 import SubKriteriaList from "@/pages/sub-kriteria/list";
-import AlternatifList from "@/pages/users-management/alternatifList";
+import AlternatifList from "@/pages/alternatif/alternatifList";
 import AdminList from "./pages/admin/adminList";
 
 function App() {
