+++ src/components/alternatif/alternatifTableActions.jsx
@@ -15,7 +15,7 @@ export default function AlternatifTableActions() {
       />
       <div className="flex justify-center">
         <button
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
+          className="bg-yellow-500 hover:bg-yellow-700 text-white font-bold py-2 px-4 rounded"
           onClick={() => setIsModalOpen(true)}
         >
           Edit
